app-cli/requirements.txt
msal==1.34.0
plotly==6.5.1
pydantic==2.12.5
pyinstaller==6.17.0
pyyaml==6.0.3
requests==2.32.5
ruff==0.14.11
typer==0.21.1
